fix signed/unsigned comparison with cast
darcs-hash:20070212205557-c9803-dbe4d7bc9536c526964bb36d229d568ab91b44f1.gz

+++ agent/test-poll.c
@@ -64,7 +64,7 @@ main (void)
 
   readable = nice_agent_poll_read (agent, fds, NULL, NULL);
   g_assert (g_slist_length (readable) == 1);
-  g_assert (GPOINTER_TO_UINT (readable->data) == pipe_fds[0]);
+  g_assert (GPOINTER_TO_UINT (readable->data) == (guint) pipe_fds[0]);
   g_slist_free (readable);
   g_assert (cb_called == FALSE);
